Reverted VisualEditor and ConfirmEdit to mw-131 stock
