+++ extensions/ConfirmEdit/ReCaptchaNoCaptcha/ReCaptchaNoCaptcha.class.php
@@ -80,14 +80,9 @@ HTML;
 	 * @return array
 	 */
 	protected function getCaptchaParamsFromRequest( WebRequest $request ) {
-		// ReCaptchaNoCaptcha combines captcha ID + solution into a single value
+		$index = 'not used'; // ReCaptchaNoCaptcha combines captcha ID + solution into a single value
 		// API is hardwired to return captchaWord, so use that if the standard isempty
-		// "captchaWord" is sent as "captchaword" by visual editor
-		$index = 'not used';
-		$response = $request->getVal( 'g-recaptcha-response',
-						$request->getVal( 'captchaWord',
-							$request->getVal( 'captchaword' )
-						) );
+		$response = $request->getVal( 'g-recaptcha-response', $request->getVal( 'captchaWord' ) );
 		return [ $index, $response ];
 	}
