+++ extensions/ConfirmEdit/ReCaptchaNoCaptcha/extension.json
@@ -7,7 +7,6 @@
 	},
 	"AutoloadClasses": {
 		"ReCaptchaNoCaptcha": "ReCaptchaNoCaptcha.class.php",
-		"ReCaptchaNoCaptchaHooks": "includes/ReCaptchaNoCaptchaHooks.php",
 		"HTMLReCaptchaNoCaptchaField": "HTMLReCaptchaNoCaptchaField.php",
 		"ReCaptchaNoCaptchaAuthenticationRequest": "ReCaptchaNoCaptchaAuthenticationRequest.php"
 	},
@@ -17,20 +16,5 @@
 		"ReCaptchaSecretKey": "",
 		"ReCaptchaSendRemoteIP": false
 	},
-	"ResourceFileModulePaths": {
-		"localBasePath": "resources",
-		"remoteExtPath": "ConfirmEdit/ReCaptchaNoCaptcha/resources"
-	},
-	"ResourceModules": {
-		"ext.confirmEdit.reCaptchaNoCaptcha.visualEditor": {
-			"scripts": "ve-confirmedit-reCaptchaNoCaptcha/ve.init.mw.NoCaptchaReCaptchaSaveErrorHandler.js"
-		}
-	},
-	"Hooks": {
-		"ResourceLoaderGetConfigVars": "ReCaptchaNoCaptchaHooks::onResourceLoaderGetConfigVars"
-	},
-	"VisualEditorPluginModules": [
-		"ext.confirmEdit.reCaptchaNoCaptcha.visualEditor"
-	],
 	"manifest_version": 1
 }
