+++ extensions/ConfirmEdit/gitinfo.json
@@ -1,1 +1,1 @@
-{"headSHA1": "b52469cf5716165dcf55e0db103d45af045b2acc\n", "head": "b52469cf5716165dcf55e0db103d45af045b2acc\n", "remoteURL": "https://gerrit.wikimedia.org/r/mediawiki/extensions/ConfirmEdit", "branch": "b52469cf5716165dcf55e0db103d45af045b2acc\n", "headCommitDate": "1525494984"}
+{"headCommitDate": "1525494984", "branch": "b52469cf5716165dcf55e0db103d45af045b2acc\n", "head": "b52469cf5716165dcf55e0db103d45af045b2acc\n", "headSHA1": "b52469cf5716165dcf55e0db103d45af045b2acc\n", "remoteURL": "https://gerrit.wikimedia.org/r/mediawiki/extensions/ConfirmEdit"}
