--- extensions/ConfirmEdit/phpcs.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0"?>
-<ruleset>
-	<rule ref="vendor/mediawiki/mediawiki-codesniffer/MediaWiki"/>
-	<file>.</file>
-	<arg name="extensions" value="php,php5,inc"/>
-	<arg name="encoding" value="utf8"/>
-	<exclude-pattern>ReCaptcha/recaptchalib.php</exclude-pattern>
-	<exclude-pattern>vendor</exclude-pattern>
-</ruleset>
