+++ extensions/VisualEditor/gitinfo.json
@@ -1,1 +1,1 @@
-{"headCommitDate": "1567576450", "head": "c3c9140b37d0b6d2d4a4ca786140ce747346a709\n", "headSHA1": "c3c9140b37d0b6d2d4a4ca786140ce747346a709\n", "branch": "c3c9140b37d0b6d2d4a4ca786140ce747346a709\n", "remoteURL": "https://gerrit-replica.wikimedia.org/r/mediawiki/extensions/VisualEditor"}
+{"branch": "c3c9140b37d0b6d2d4a4ca786140ce747346a709\n", "head": "c3c9140b37d0b6d2d4a4ca786140ce747346a709\n", "remoteURL": "https://gerrit-replica.wikimedia.org/r/mediawiki/extensions/VisualEditor", "headSHA1": "c3c9140b37d0b6d2d4a4ca786140ce747346a709\n", "headCommitDate": "1567576450"}
